+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md).
 
 ## [Unreleased]
+### Added
+- `check-postgres-query.rb`: Add `-r`, `--regex-pattern` to match query result against (@jindraj)
+
 ### Changes
 - Updated development dependency to bundler ~> 2.1
 - Updated development dependency to rake ~> 13.0